+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/anthropic
 
+## 3.0.48
+
+### Patch Changes
+
+- 2164cdf: feat(anthropic): add the new code_execution tool
+
+## 3.0.47
+
+### Patch Changes
+
+- 17978c6: Pass `cacheControl` provider option as top-level `cache_control` in Anthropic API request body to support automatic caching.
+
 ## 3.0.46
 
 ### Patch Changes